Added auth module and moved previous authentication store to the module
