+++ src/router.js
@@ -5,7 +5,7 @@ import Home from "./views/Home.vue";
 import Events from "./views/Events.vue";
 import Staff from "./views/Staff.vue";
 import Help from "./views/Help.vue";
-import store from "./store";
+import store from "./state/store";
 import Calendar from "./views/Calendar";
 
 Vue.use(Router);
